my_agent/requirements.txt
langgraph
langchain_openai
langchain_core
langchain_google_genai
openai
python-dotenv 